Exception Handlers have been added

+++ src/main/java/com/ecommerce/sb_ecom/service/CategoryService.java
@@ -9,7 +9,7 @@ public interface CategoryService {
     public List<Category> getAllCategories();
     public String addCategory(Category category);
 
-    public String deleteCategory(int categoryId);
+    public String deleteCategory(Long categoryId);
 
     public Category updateCategory(Category category, Long categoryId);
 }
